Stock prices CRUD done

+++ src/Core/Models/SimulationStep.cs
@@ -1,3 +1,3 @@
 ﻿namespace Core.Models;
 
-public readonly record struct SimulationStep(int Step);
+public readonly record struct SimulationStep(long Step);
